+++ hypothesis-python/tests/cover/test_given_error_conditions.py
@@ -19,20 +19,10 @@ from __future__ import division, print_function, absolute_import
 
 import pytest
 
-from hypothesis import HealthCheck, given, infer, assume, reject, settings
+from hypothesis import given, infer, assume, reject
 from hypothesis.errors import Unsatisfiable, InvalidArgument
 from tests.common.utils import fails_with, validate_deprecation
 from hypothesis.strategies import booleans, integers
-
-
-def test_raises_unsatisfiable_if_all_false():
-    @given(integers())
-    @settings(max_examples=50, suppress_health_check=HealthCheck.all())
-    def test_assume_false(x):
-        reject()
-
-    with pytest.raises(Unsatisfiable):
-        test_assume_false()
 
 
 def test_raises_unsatisfiable_if_all_false_in_finite_set():
