+++ hypothesis-python/tests/nocover/test_simple_strings.py
@@ -1,0 +1,29 @@
+# coding=utf-8
+#
+# This file is part of Hypothesis, which may be found at
+# https://github.com/HypothesisWorks/hypothesis-python
+#
+# Most of this work is copyright (C) 2013-2018 David R. MacIver
+# (david@drmaciver.com), but it contains contributions by others. See
+# CONTRIBUTING.rst for a full list of people who may hold copyright, and
+# consult the git log if you need to determine who owns an individual
+# contribution.
+#
+# This Source Code Form is subject to the terms of the Mozilla Public License,
+# v. 2.0. If a copy of the MPL was not distributed with this file, You can
+# obtain one at http://mozilla.org/MPL/2.0/.
+#
+# END HEADER
+
+from __future__ import division, print_function, absolute_import
+
+import unicodedata
+
+from hypothesis import given, settings
+from hypothesis.strategies import text
+
+
+@given(text(min_size=1, max_size=1))
+@settings(max_examples=2000)
+def test_does_not_generate_surrogates(t):
+    assert unicodedata.category(t) != u'Cs'
